refactor: 🔧 modularize init.lua configuration
